rename hosted web suite catalog module
