+++ packages/test-cases/src/index.ts
@@ -4,4 +4,4 @@ export * from "./apps/forum-lite.js";
 export * from "./apps/repo-lite.js";
 export * from "./apps/wiki-lite.js";
 export * from "./apps/notes-lite.js";
-export * from "./suites/hosted-web-v2.js";
+export * from "./suites/hosted-web.js";
